Sprint2

+++ Application/src/main/java/com/kenzie/appserver/service/RecipeService.java
@@ -8,7 +8,9 @@ import com.kenzie.appserver.repositories.RecipeRepository;
 import com.kenzie.appserver.service.model.Recipe;
 import com.kenzie.capstone.service.client.LambdaRecipeServiceClient;
 import com.kenzie.capstone.service.model.*;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -28,6 +30,10 @@ public class RecipeService {
 
     public Recipe findById(String recipeId) {
 
+        if (recipeId == null || recipeId.length() == 0) {
+            throw new IllegalArgumentException("Recipe Id is not valid");
+        }
+
         //get it from the cache if it exists
         Recipe cachedRecipe = cache.get(recipeId);
         if (cachedRecipe != null){
@@ -37,47 +43,30 @@ public class RecipeService {
         // getting data from the lambda
         RecipeData recipeFromLambda = lambdaRecipeServiceClient.getRecipeData(recipeId);
 
-        // getting data from the local repository
-//        Recipe dataFromDynamo = recipeRepository
-//                .findById(recipeId)
-//                .map(recipe -> new Recipe(recipe.getRecipeId(), recipe.getTitle(), recipe.getIngredients(),
-//                        recipe.getSteps(), recipe.getIsGlutenFree(), recipe.getIsDairyFree(), recipe.getIsEggFree(),
-//                        recipe.getIsVegetarian(), recipe.getIsVegan(), recipe.getRatings()))
-//                .orElse(null);
-
-        //return dataFromDynamo;
-        //return null;
-
         Recipe recipe = new Recipe(recipeFromLambda.getRecipeId(),
                 recipeFromLambda.getTitle(), recipeFromLambda.getIngredients(),
                 recipeFromLambda.getSteps(), recipeFromLambda.isGlutenFree(),
                 recipeFromLambda.isDairyFree(), recipeFromLambda.isEggFree(),
                 recipeFromLambda.isVegetarian(), recipeFromLambda.isVegan(), recipeFromLambda.getRatings());
+
         return recipe;
     }
 
     public Recipe addNewRecipe(RecipeCreateRequest recipeCreateRequest) {
+
+        if (recipeCreateRequest.getTitle() == null || recipeCreateRequest.getTitle().length() == 0) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid Recipe");
+        }
+
         //convert RecipeCreateRequest to Recipe Request
         List<Integer> ratings = new ArrayList<> ();
         RecipeRequest recipeRequest = new RecipeRequest(recipeCreateRequest.getTitle(),
                 recipeCreateRequest.getIngredients(), recipeCreateRequest.getSteps(), recipeCreateRequest.isGlutenFree(),
                 recipeCreateRequest.isDairyFree(), recipeCreateRequest.isEggFree(), recipeCreateRequest.isVegetarian(),
                 recipeCreateRequest.isVegan(), ratings);
+
         // sending data to Lambda
         RecipeResponse recipeFromLambda = lambdaRecipeServiceClient.setRecipeData(recipeRequest);
-
-        // sending data to the local repository
-//        RecipeRecord recipeRecord = new RecipeRecord();
-//        recipeRecord.setRecipeId(recipeFromLambda.getRecipeId());
-//        recipeRecord.setTitle(recipeFromLambda.getTitle());
-//        recipeRecord.setIngredients(recipeFromLambda.getIngredients());
-//        recipeRecord.setSteps(recipeFromLambda.getSteps());
-//        recipeRecord.setIsGlutenFree(recipeFromLambda.isGlutenFree());
-//        recipeRecord.setIsDairyFree(recipeFromLambda.isDairyFree());
-//        recipeRecord.setIsEggFree(recipeFromLambda.isEggFree());
-//        recipeRecord.setIsVegetarian(recipeFromLambda.isVegetarian());
-//        recipeRecord.setIsVegan(recipeFromLambda.isVegan());
-//        recipeRepository.save(recipeRecord);
 
         return new Recipe(recipeFromLambda.getRecipeId(),
                             recipeFromLambda.getTitle(), recipeFromLambda.getIngredients(),
@@ -87,6 +76,10 @@ public class RecipeService {
     }
 
     public List<Recipe> findByDietaryRestriction(DietaryRestrictionInfoRequest dietaryRestrictionInfoRequest){
+
+        if (dietaryRestrictionInfoRequest == null) {
+            throw new IllegalArgumentException("Invalid Dietary Restrictions");
+        }
 
         //getting it from the lambda
         DietaryRestrictionData data = dietaryRestrictionInfoRequestToData(dietaryRestrictionInfoRequest);
@@ -105,6 +98,15 @@ public class RecipeService {
     }
 
     public Recipe updateRecipe(RecipeUpdateRequest recipeUpdateRequest){
+
+        if (recipeUpdateRequest.getRecipeId() == null || recipeUpdateRequest.getRecipeId().length() == 0) {
+            throw new IllegalArgumentException("Invalid Recipe Id");
+        }
+
+        if (recipeUpdateRequest.getNewRating() == null ||
+                recipeUpdateRequest.getNewRating() < 0 || recipeUpdateRequest.getNewRating() > 5) {
+            throw new IllegalArgumentException("Invalid Rating");
+        }
 
         Recipe recipe = cache.get(recipeUpdateRequest.getRecipeId());
 
@@ -138,13 +140,19 @@ public class RecipeService {
 
     public void deleteRecipe(String recipeId) {
 
+        if (recipeId == null || recipeId.length() == 0) {
+            throw new IllegalArgumentException("Recipe Id is not valid");
+        }
+
         lambdaRecipeServiceClient.deleteById(recipeId);
+
         cache.evict(recipeId);
 
     }
     //helper methods
 
     private Recipe recipeDataToRecipe(RecipeData data){
+
         Recipe recipe = new Recipe(data.getRecipeId(),
                 data.getTitle(), data.getIngredients(),
                 data.getSteps(), data.isGlutenFree(),
@@ -155,12 +163,14 @@ public class RecipeService {
     }
 
     private DietaryRestrictionData dietaryRestrictionInfoRequestToData(DietaryRestrictionInfoRequest request){
+
         DietaryRestrictionData data = new DietaryRestrictionData();
         data.setIsGlutenFree(request.isGlutenFree());
         data.setIsDairyFree(request.isDairyFree());
         data.setIsEggFree(request.isEggFree());
         data.setIsVegetarian(request.isVegetarian());
         data.setIsVegan(request.isVegan());
+
         return data;
     }
